+++ src/commands/hardis/project/deploy/quick.ts
@@ -11,6 +11,36 @@ export default class ProjectDeployStart extends SfCommand<any> {
 [![Assisted solving of Salesforce deployments errors](https://github.com/hardisgroupcom/sfdx-hardis/raw/main/docs/assets/images/article-deployment-errors.jpg)](https://nicolas.vuillamy.fr/assisted-solving-of-salesforce-deployments-errors-47f3666a9ed0)
 
 [See documentation of Salesforce command](https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_project_commands_unified.htm#cli_reference_project_deploy_quick_unified)
+
+### Deployment pre or post commands
+
+You can define command lines to run before or after a deployment, with parameters:
+
+- **id**: Unique Id for the command
+- **label**: Human readable label for the command
+- **skipIfError**: If defined to "true", the post-command won't be run if there is a deployment failure
+- **context**: Defines the context where the command will be run. Can be **all** (default), **check-deployment-only** or **process-deployment-only**
+
+If the commands are not the same depending on the target org, you can define them into **config/branches/.sfdx-hardis-BRANCHNAME.yml** instead of root **config/.sfdx-hardis.yml**
+
+Example:
+
+\`\`\`yaml
+commandsPreDeploy:
+  - id: knowledgeUnassign
+    label: Remove KnowledgeUser right to the user who has it
+    command: sf data update record --sobject User --where "UserPermissionsKnowledgeUser='true'" --values "UserPermissionsKnowledgeUser='false'" --json
+  - id: knowledgeAssign
+    label: Assign Knowledge user to the deployment user
+    command: sf data update record --sobject User --where "Username='deploy.github@myclient.com'" --values "UserPermissionsKnowledgeUser='true'" --json
+commandsPostDeploy:
+  - id: knowledgeUnassign
+    label: Remove KnowledgeUser right to the user who has it
+    command: sf data update record --sobject User --where "UserPermissionsKnowledgeUser='true'" --values "UserPermissionsKnowledgeUser='false'" --json
+  - id: knowledgeAssign
+    label: Assign Knowledge user to desired username
+    command: sf data update record --sobject User --where "Username='admin-yser@myclient.com'" --values "UserPermissionsKnowledgeUser='true'" --json
+\`\`\`
 `;
 
   public static flags: any = {
@@ -52,7 +82,7 @@ export default class ProjectDeployStart extends SfCommand<any> {
   public async run(): Promise<AnyJson> {
     const { flags } = await this.parse(ProjectDeployStart);
     // Run pre deployment commands if defined
-    await executePrePostCommands('commandsPreDeploy', true);
+    await executePrePostCommands('commandsPreDeploy', { success: true, checkOnly: false });
     const result = await wrapSfdxCoreCommand("sf project deploy start", this.argv, this, flags.debug);
     // Check org coverage if requested
     if (flags['coverage-formatters'] && result.stdout) {
@@ -68,7 +98,7 @@ export default class ProjectDeployStart extends SfCommand<any> {
       }
     }
     // Run post deployment commands if defined
-    await executePrePostCommands('commandsPostDeploy', process.exitCode === 0);
+    await executePrePostCommands('commandsPostDeploy', { success: process.exitCode === 0, checkOnly: false });
     await GitProvider.managePostPullRequestComment();
     return result;
   }