+++ CHANGELOG.md
@@ -102,13 +102,18 @@ We made many tests but risk zero do not exist, so if you see any bug, please rep
 
 ### New Features / Enhancements
 
+- commandsPreDeploy and commandsPostDeploy: Allow to define a running context for a command: **all** (default), **check-deployment-only** or **process-deployment-only****process-deployment-only**
 - New command **hardis:project:deploy:simulate** to validate the deployment of a single metadata (used by VsCode extension)
 - New command **hardis:org:diagnose:releaseupdates** to check for org Release Updates from Monitoring or locally
 - New command **hardis:misc:purge-references** to partially automate the cleaning of related dependencies when you need to delete a field, or change its type (for example from master detail to lookup)
+- New command **hardis:project:clean:sensitive-metadatas** to mask sensitive metadatas from git repo (ex: Certificate content)
 - **hardis:work:save** and **hardis:project:deploy:sources:dx**: Improve runtime performances thanks to internalization of sfdx-essentials commands
 - **hardis:work:new**: Allow to add labels in property `availableTargetBranches`, using a comma. For examples, `- integration,Choose this branch if you are on the BUILD side of the project !`
 - **hardis:work:new**: Add current default org in the choices when prompting which org to use
 - Rename command **hardis:project:deploy:source:dx** into **hardis:project:deploy:smart**
+- **hardis:project:new**
+  - Initialize autoCleanTypes with **destructivechanges**, **flowPositions** and **minimizeProfiles**
+  - Initialize package-no-overwrite.xml with Certificate metadata. (certificates must be uploaded manually)
 - **hardis:org:files:export**: Improve display with spinner
 - **hardis:org:purge:flow**: If FlowInterview records are preventing Flow Versions to be deleted, prompt user to delete Flow Interviews before trying again to delete Flow Versions
 - **hardis:project:generate:gitdelta**: Add option to generate package.xml related to a single commit
@@ -125,7 +130,8 @@ We made many tests but risk zero do not exist, so if you see any bug, please rep
   - Customer Portal: createdcustomersuccessuser
 - AI Deployment assistant
   - Add error `Change Matching Rule`
-  - Improve documentation by adding examples of errors, and a standalone page for each tip
+- Git Providers
+  - On Pull Requests / Merge Requests comments, add hyperlinks to errors documentation URL
 
 ### Fixes
 
@@ -137,6 +143,8 @@ We made many tests but risk zero do not exist, so if you see any bug, please rep
 
 - Reorganize README content
   - Add link to Dreamforce 24 session
+- Deployment assistant: Improve documentation by adding examples of errors, and a standalone page for each tip
+- Factorize the definition of DOC_ROOT_URL https://sfdx-hardis.cloudity.com
 
 ## [4.53.0] 2024-08-20